Add parallel bindings, cleanup examples (#9)
* Make examples consistent for each language

* Add parallel bindings for C, with test

* Fix case of fortran test file name

* Add parallel Fortran infrastructure

* Add parallel Fortran demo

* Cleanup CMake for MPI builds

* Make setup functions idempotent

+++ test/fullexample.py
@@ -60,7 +60,7 @@ class TestHelpme(unittest.TestCase):
 
         energy = 0
         forces = np.zeros((6,3),dtype=np.float32)
-        virial = np.zeros((1,6),dtype=np.float64)
+        virial = np.zeros((1,6),dtype=np.float32)
 
         pmeF = pme.PMEInstanceF()
         pmeF.setup(1, 0.3, 5, 32, 32, 32, 332.0716, 1)
